requirements.txt
Pillow
matplotlib
numpy
opencv-python